requirements.txt
python-nmap
rich